Backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
firebase-admin>=6.4,<7.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
hypothesis>=6.100,<7.0
bcrypt>=4.1,<5.0
